feat: setup metrics monitoring
Deploy the kube-prometheus-stack helm chart. I've disabled alert-manager
for now, as I'm not yet familiar with it. However, I plan to deploy it
in the near future. Additionally, I've disabled metrics for the
controller-manager, scheduler, and kube-proxy, as these require changes
to the k3s configuration that I haven't tested yet.