ci: setup nixos config & workflow for node `base`
Since this new node is within my home network, I had to set up a GitHub runner
on it and modify the workflow to use the self-hosted runner. This means the
NixOS configuration will be built directly on this node.

I'm planning to use this spare laptop as a K3S worker node.